refactor(contacts): colocate sanctioned address banner
